refactor: delete ValidationOperatorAnonymizer
