+++ great_expectations/core/usage_statistics/usage_statistics.py
@@ -63,6 +63,9 @@ class UsageStatisticsHandler:
         # the risk of cyclic import issues. If these anonymizers have been imported at any earlier point
         # in the program's lifetime, retrieval of the import will be O(1) and not impact performance.
 
+        from great_expectations.core.usage_statistics.anonymizers.anonymizer import (
+            Anonymizer,
+        )
         from great_expectations.core.usage_statistics.anonymizers.batch_anonymizer import (
             BatchAnonymizer,
         )
@@ -90,16 +93,11 @@ class UsageStatisticsHandler:
         from great_expectations.core.usage_statistics.anonymizers.store_anonymizer import (
             StoreAnonymizer,
         )
-        from great_expectations.core.usage_statistics.anonymizers.validation_operator_anonymizer import (
-            ValidationOperatorAnonymizer,
-        )
 
+        self._anonymizer = Anonymizer(data_context_id)
         self._datasource_anonymizer = DatasourceAnonymizer(data_context_id)
         self._execution_engine_anonymizer = ExecutionEngineAnonymizer(data_context_id)
         self._store_anonymizer = StoreAnonymizer(data_context_id)
-        self._validation_operator_anonymizer = ValidationOperatorAnonymizer(
-            data_context_id
-        )
         self._data_docs_sites_anonymizer = DataDocsSiteAnonymizer(data_context_id)
         self._batch_request_anonymizer = BatchRequestAnonymizer(data_context_id)
         self._batch_anonymizer = BatchAnonymizer(data_context_id)
@@ -176,7 +174,7 @@ class UsageStatisticsHandler:
                 for store_name, store_obj in self._data_context.stores.items()
             ],
             "anonymized_validation_operators": [
-                self._validation_operator_anonymizer.anonymize_validation_operator_info(
+                self._anonymizer.anonymize_validation_operator_info(
                     validation_operator_name=validation_operator_name,
                     validation_operator_obj=validation_operator_obj,
                 )
